# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: cloudflared-manager API
  description: |
    无头（headless）cloudflared **TOKEN 模式** 隧道管理器。守护进程 `cfdmgrd` 把每个
    cloudflared 连接器（connector）放进**独立的外部子进程**（`cloudflared tunnel run --token …`），
    通过 HTTP/JSON 暴露隧道 CRUD、生命周期、历史曲线、告警、日志、导入导出、系统监控、
    cloudflared 二进制版本管理等管理面。前端 SPA 已 embed 进二进制，生产环境同域。

    关键架构事实（落到 API 字段命名、状态码、错误形态都需要先看一眼）：

    - **外部子进程模型**：每个运行中的隧道都是 `cloudflared tunnel run` 子进程，由
      `internal/process.Worker` 监管。父进程通过 `--token`（环境变量 `TUNNEL_TOKEN`）注入
      连接器令牌；ingress / public-hostname / origin 等参数全部在 Cloudflare Zero Trust
      控制台维护，**不在本模型内**。
    - **reload = 重启**：cloudflared 没有针对单连接器参数的 in-place 热重载语义，
      `POST .../reload` 的实现就是 `stop()` 再 `start()`。
    - **token 高度敏感**：所有返回 `ConfigEnvelope` 的接口里 `config.token` 一律被剥除（空串），
      仅用布尔 `has_token` 表明是否已存储；掩码预览经 `GET /configs/{id}/token` 获取。
      **唯一例外** 是 `GET /configs/{id}/raw`（原样回写磁盘 YAML，按设计包含 token，属敏感导出）。
    - **流量曲线语义特殊**：cloudflared **没有按隧道的字节计数器**。
      `GET /metrics/{id}/traffic` 返回的 `points[].in` = 请求数增量、`out` = 错误数增量、
      `conns` = HA 连接数 gauge（**不是字节**），均来自抓取 cloudflared 自带的 `--metrics`
      Prometheus 端点。
    - **配置数据模型**：业务层（请求/响应里的 `config` 字段）= `TunnelConfigV1`，**全 camelCase**
      （edge/reliability/logging/identity/advancedEnvOverrides/binaryVersion）。
    - **快照 / 事件 / 告警 / 流量 / 系统监控**：`Snapshot`、`/metrics/*`、`/alerts/*`、`/system/*`、
      WebSocket `/events` 帧统一走 **snake_case**（id / inst_id / started_at / metrics_port …）。
...

...

...
  version: 2.0.0
servers:
...
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/cf/accounts:
    get:
      tags: [cloudflare]
      summary: 列出 Cloudflare 账号（脱敏）
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/CFAccountView" } }
    post:
      tags: [cloudflare]
      summary: 新建账号（自动校验并探测 account_id）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CFAccountInput" }
      responses:
        "201":
          description: 已创建（含校验结果）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CFAccountResp" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/v1/cf/accounts/{aid}:
    parameters:
      - { name: aid, in: path, required: true, schema: { type: string } }
    get:
      tags: [cloudflare]
      summary: 账号详情（脱敏）
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/CFAccountView" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [cloudflare]
      summary: 改名 / 换凭证（空 secret = 保留）并重新校验
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CFAccountInput" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/CFAccountResp" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [cloudflare]
      summary: 删除账号
      responses:
        "204": { description: 已删除 }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/cf/accounts/{aid}/verify:
    parameters:
      - { name: aid, in: path, required: true, schema: { type: string } }
    post:
      tags: [cloudflare]
      summary: 重新校验凭证、刷新状态与 account_id
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/CFAccountResp" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/cf/accounts/{aid}/cf-accounts:
    parameters:
      - { name: aid, in: path, required: true, schema: { type: string } }
    get:
      tags: [cloudflare]
      summary: 列出该凭证可见的 Cloudflare 账号（多账号选择 account_id）
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/CFAccount" } }
        "502": { description: Cloudflare 上游错误, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/cf/accounts/{aid}/tunnels:
    parameters:
      - { name: aid, in: path, required: true, schema: { type: string } }
    get:
      tags: [cloudflare]
      summary: 列出远端隧道
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/CFTunnel" } }
        "409": { $ref: "#/components/responses/Conflict", description: 账号尚未确定 account_id }
    post:
      tags: [cloudflare]
      summary: 新建远端隧道
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                config_src: { type: string, enum: [cloudflare, local], default: cloudflare }
      responses:
        "201": { description: 已创建, content: { application/json: { schema: { $ref: "#/components/schemas/CFTunnel" } } } }
  /api/v1/cf/accounts/{aid}/tunnels/{tid}:
    parameters:
      - { name: aid, in: path, required: true, schema: { type: string } }
      - { name: tid, in: path, required: true, schema: { type: string } }
    get:
      tags: [cloudflare]
      summary: 获取远端隧道
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/CFTunnel" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [cloudflare]
      summary: 重命名远端隧道
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [name], properties: { name: { type: string } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/CFTunnel" } } } }
    delete:
      tags: [cloudflare]
      summary: 删除远端隧道
      responses:
        "204": { description: 已删除 }
  /api/v1/cf/accounts/{aid}/tunnels/{tid}/token:
    parameters:
      - { name: aid, in: path, required: true, schema: { type: string } }
      - { name: tid, in: path, required: true, schema: { type: string } }
    get:
      tags: [cloudflare]
      summary: 获取远端隧道 connector token（敏感）
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: object, properties: { token: { type: string } } } } } }
  /api/v1/cf/accounts/{aid}/tunnels/{tid}/configurations:
    parameters:
      - { name: aid, in: path, required: true, schema: { type: string } }
      - { name: tid, in: path, required: true, schema: { type: string } }
    get:
      tags: [cloudflare]
      summary: 获取隧道远端配置（ingress / originRequest）
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/CFConfigurationResult" } } } }
    put:
      tags: [cloudflare]
      summary: 整体替换隧道配置（CF 仅支持全量替换）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [config]
              properties:
                config: { $ref: "#/components/schemas/CFTunnelConfig" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/CFConfigurationResult" } } } }
  /api/v1/cf/accounts/{aid}/tunnels/{tid}/connections:
    parameters:
      - { name: aid, in: path, required: true, schema: { type: string } }
      - { name: tid, in: path, required: true, schema: { type: string } }
    get:
      tags: [cloudflare]
      summary: 列出连接器
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, properties: { items: { type: array, items: { $ref: "#/components/schemas/CFConnector" } } } }
    delete:
      tags: [cloudflare]
      summary: 清理连接（client_id 省略=清理全部）
      parameters:
        - { name: client_id, in: query, required: false, schema: { type: string } }
      responses:
        "204": { description: 已清理 }
  /api/v1/cf/accounts/{aid}/zones:
    parameters:
      - { name: aid, in: path, required: true, schema: { type: string } }
      - { name: name, in: query, required: false, schema: { type: string }, description: 精确域名过滤 }
    get:
      tags: [cloudflare]
      summary: 列出 DNS zones
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, properties: { items: { type: array, items: { $ref: "#/components/schemas/CFZone" } } } }
  /api/v1/cf/accounts/{aid}/zones/{zid}/dns_records:
    parameters:
      - { name: aid, in: path, required: true, schema: { type: string } }
      - { name: zid, in: path, required: true, schema: { type: string } }
    get:
      tags: [cloudflare]
      summary: 列出 DNS 记录
      parameters:
        - { name: name, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, properties: { items: { type: array, items: { $ref: "#/components/schemas/CFDNSRecord" } } } }
    post:
      tags: [cloudflare]
      summary: 新建 DNS 记录
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CFDNSRecord" }
      responses:
        "201": { description: 已创建, content: { application/json: { schema: { $ref: "#/components/schemas/CFDNSRecord" } } } }
  /api/v1/cf/accounts/{aid}/zones/{zid}/dns_records/{rid}:
    parameters:
      - { name: aid, in: path, required: true, schema: { type: string } }
      - { name: zid, in: path, required: true, schema: { type: string } }
      - { name: rid, in: path, required: true, schema: { type: string } }
    put:
      tags: [cloudflare]
      summary: 替换 DNS 记录
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CFDNSRecord" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/CFDNSRecord" } } } }
    delete:
      tags: [cloudflare]
      summary: 删除 DNS 记录
      responses:
        "204": { description: 已删除 }
  /api/v1/configs/{id}/cf/token-info:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [cloudflare]
      summary: 解码本地实例 token → account_tag / tunnel_id（无 secret）
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CFTokenInfo" }
  /api/v1/configs/{id}/cf/binding:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [cloudflare]
      summary: 当前实例↔隧道绑定 + 远端隧道概要
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/CFBinding" } } } }
    put:
      tags: [cloudflare]
      summary: 关联实例到隧道（归属校验）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [account_id]
              properties:
                account_id: { type: string, description: 本地 cfaccount id }
                tunnel_id: { type: string, description: 缺省时用 token 解出的 tunnel_id }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/CFBinding" } } } }
        "400": { $ref: "#/components/responses/BadRequest", description: 隧道不属于此账号 / 无法确定 tunnel_id }
    delete:
      tags: [cloudflare]
      summary: 解绑
      responses:
        "204": { description: 已解绑 }
  /api/v1/configs/{id}/cf/public-hostnames:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [cloudflare]
      summary: 聚合实例的公共主机名（ingress + 各自 DNS 状态）
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/CFPublicHostnameList" } } } }
        "409": { $ref: "#/components/responses/Conflict", description: 实例未关联账号 }
    post:
      tags: [cloudflare]
      summary: 新增公共主机名（插入 ingress + 建/改代理 CNAME）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CFPublicHostnameInput" }
      responses:
        "201": { description: 已创建, content: { application/json: { schema: { type: object } } } }
  /api/v1/configs/{id}/cf/public-hostnames/{index}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
      - { name: index, in: path, required: true, schema: { type: integer }, description: ingress 规则下标（不能指向兜底规则） }
    put:
      tags: [cloudflare]
      summary: 修改第 index 条公共主机名（含 DNS 同步）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CFPublicHostnameInput" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: object } } } }
    delete:
      tags: [cloudflare]
      summary: 删除第 index 条公共主机名
      parameters:
        - { name: delete_dns, in: query, required: false, schema: { type: string, enum: ["true"] }, description: 同时删除代理 CNAME }
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: object } } } }
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "Authorization: Bearer <CFDM_API_TOKEN>。WebSocket 接口可改用 ?token= 查询参数。"
  responses:
    BadRequest:
      description: 请求非法
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: 状态冲突
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      description: 统一错误信封
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: 机器可读错误码
              enum:
                - bad_request
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - validation_failed
                - internal_error
                - config_not_found
                - config_already_exists
                - invalid_state
                - upstream_failure
            message: { type: string }
            details: { type: object, additionalProperties: true }
...
    CFAccountInput:
      type: object
      description: "新建/更新账号入参。更新时空 secret = 保留原值。"
      required: [name, auth_type]
      properties:
        name: { type: string }
        auth_type: { type: string, enum: [token, key] }
        token: { type: string, description: "API Token（auth_type=token）" }
        email: { type: string, description: "账户邮箱（auth_type=key）" }
        api_key: { type: string, description: "Global API Key（auth_type=key）" }
        account_id: { type: string, description: "可选，显式指定 Cloudflare account id；留空自动探测" }
    CFAccountView:
      type: object
      description: "脱敏账号视图（snake_case）。secret 永不回传，仅 has_token/has_key。"
      properties:
        id: { type: string, description: 本地账号 id }
        name: { type: string }
        auth_type: { type: string, enum: [token, key] }
        account_id: { type: string, description: Cloudflare account id }
        account_name: { type: string }
        email: { type: string }
        has_token: { type: boolean }
        has_key: { type: boolean }
        status: { type: string, enum: [unverified, active, invalid] }
        last_verified_at: { type: string }
        created_at: { type: string }
        updated_at: { type: string }
    CFAccount:
      type: object
      description: "Cloudflare 账号（GET /accounts 元素）"
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string }
    CFAccountVerify:
      type: object
      properties:
        ok: { type: boolean }
        error: { type: string }
        accounts: { type: array, items: { $ref: "#/components/schemas/CFAccount" } }
    CFAccountResp:
      type: object
      properties:
        account: { $ref: "#/components/schemas/CFAccountView" }
        verify: { $ref: "#/components/schemas/CFAccountVerify" }
    CFTunnel:
      type: object
      description: "远端隧道（cfd_tunnel，snake_case）"
      properties:
        id: { type: string }
        account_tag: { type: string }
        name: { type: string }
        created_at: { type: string }
        deleted_at: { type: string }
        conns_active_at: { type: string }
        conns_inactive_at: { type: string }
        status: { type: string, enum: [inactive, degraded, healthy, down] }
        tun_type: { type: string }
        config_src: { type: string, enum: [local, cloudflare] }
        remote_config: { type: boolean }
        connections: {}
    CFIngressRule:
      type: object
      description: "ingress 规则（camelCase，原样直传 cloudflared）"
      properties:
        hostname: { type: string }
        service: { type: string, description: "如 http://localhost:8080 / tcp://... / http_status:404" }
        path: { type: string }
        originRequest: { type: object, additionalProperties: true, description: "连接参数：connectTimeout/noTLSVerify/httpHostHeader/access{required,teamName,audTag[]}…" }
    CFTunnelConfig:
      type: object
      properties:
        ingress: { type: array, items: { $ref: "#/components/schemas/CFIngressRule" } }
        originRequest: { type: object, additionalProperties: true, description: 全局默认连接参数 }
        warp-routing: { type: object, properties: { enabled: { type: boolean } } }
    CFConfigurationResult:
      type: object
      properties:
        account_id: { type: string }
        tunnel_id: { type: string }
        version: { type: integer }
        config: { $ref: "#/components/schemas/CFTunnelConfig" }
        source: { type: string }
        created_at: { type: string }
    CFConnector:
      type: object
      properties:
        id: { type: string }
        features: { type: array, items: { type: string } }
        version: { type: string }
        arch: { type: string }
        run_at: { type: string }
        config_version: { type: integer }
        conns: {}
    CFZone:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        status: { type: string }
        paused: { type: boolean }
        account: { type: object, properties: { id: { type: string }, name: { type: string } } }
    CFDNSRecord:
      type: object
      properties:
        id: { type: string }
        type: { type: string, description: "CNAME / A / TXT …（隧道公共主机名用 CNAME）" }
        name: { type: string }
        content: { type: string, description: "隧道 CNAME 指向 <tunnel_id>.cfargotunnel.com" }
        proxied: { type: boolean }
        ttl: { type: integer, description: "1 = 自动" }
        comment: { type: string }
        proxiable: { type: boolean }
        created_on: { type: string }
        modified_on: { type: string }
    CFTokenInfo:
      type: object
      properties:
        has_token: { type: boolean }
        account_tag: { type: string }
        tunnel_id: { type: string }
        error: { type: string }
    CFBinding:
      type: object
      properties:
        bound: { type: boolean }
        account_id: { type: string }
        account_name: { type: string }
        cf_account_id: { type: string }
        tunnel_id: { type: string }
        tunnel_name: { type: string }
        account_tag: { type: string }
        token_account_tag: { type: string }
        token_tunnel_id: { type: string }
        match: { type: boolean, description: "token 解出的 tunnel_id 是否与绑定一致" }
        tunnel: { $ref: "#/components/schemas/CFTunnel" }
    CFDNSStatus:
      type: object
      properties:
        zone_id: { type: string }
        zone_name: { type: string }
        record_id: { type: string }
        content: { type: string }
        proxied: { type: boolean }
        exists: { type: boolean }
        in_sync: { type: boolean }
    CFPublicHostname:
      type: object
      properties:
        index: { type: integer }
        hostname: { type: string }
        service: { type: string }
        path: { type: string }
        origin_request: { type: object, additionalProperties: true }
        dns: { $ref: "#/components/schemas/CFDNSStatus" }
    CFPublicHostnameList:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/CFPublicHostname" } }
        tunnel_id: { type: string }
        dns_error: { type: string, description: "DNS 富集为尽力而为；权限/zone 缺失时此处给出原因，不影响主体返回" }
    CFPublicHostnameInput:
      type: object
      required: [hostname, service]
      properties:
        hostname: { type: string }
        service: { type: string }
        path: { type: string }
        origin_request: { type: object, additionalProperties: true }
        manage_dns: { type: boolean, default: true, description: "是否自动同步代理 CNAME" }